requirements.txt
aiofiles==24.1.0
aiogram==3.20.0.post0
aiohappyeyeballs==2.6.1
aiohttp==3.11.17
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==5.0.1
asyncpg==0.30.0
attrs==25.3.0
certifi==2025.1.31
cmdw==0.30
configparser==7.2.0
configset==1.54
frozenlist==1.6.0
greenlet==3.2.0
idna==3.10
magic-filter==1.0.12
make-colors==3.36
markdown-it-py==3.0.0
mdurl==0.1.2
multidict==6.4.3
propcache==0.3.1
pydantic==2.11.3
pydantic_core==2.33.1
pydebugger==0.74
Pygments==2.19.1
python-dotenv==1.1.0
rich==14.0.0
SQLAlchemy==2.0.40
typing-inspection==0.4.0
typing_extensions==4.13.2
yarl==1.20.0
